Initialize Mutex's at the start of Init process
Running GC requres these Mutex's to be initialised but the code that
initialises them (in thread.c) happens after the initialisation of the
GC.

Normally this wouldn't be a problem because GC wouldn't need to run
during the Ruby initialisation process.

Since we've added Garbage to the end of an RVALUE, we now reach the end
of a transient heap page during the startup process which triggers a GC
run - which crashes because the mutex's aren't initialised so the OS
rejects the request for the lock.

This patch extracts the mutex initialisation from Init_Thread and pulls
it right to the top of the ruby initialisation process so they're going
to be usable no matter when the GC runs.

Initially we tried to workaround this by just moving the ordering of the
initialisations in inits.c. This turned into a game of whack a mole as
there are dependencies between a lot of these items.

We chose this approach as it makes it clearer exactly which code that we
are depending on and that it needs to run first in the initialisation
queue.